+++ internal/sdkprovider/service/kafkaschema/kafka_schema_configuration.go
@@ -2,16 +2,21 @@ package kafkaschema
 
 import (
 	"context"
+	"path/filepath"
 
-	"github.com/aiven/aiven-go-client/v2"
+	avngen "github.com/aiven/go-client-codegen"
 	"github.com/aiven/go-client-codegen/handler/kafkaschemaregistry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	"golang.org/x/sync/singleflight"
 
+	"github.com/aiven/terraform-provider-aiven/internal/common"
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil"
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil/userconfig"
 )
+
+var schemaRegistryCheckGroup singleflight.Group
 
 var aivenKafkaSchemaConfigurationSchema = map[string]*schema.Schema{
 	"project":      schemautil.CommonSchemaProjectReference,
@@ -33,10 +38,10 @@ var aivenKafkaSchemaConfigurationSchema = map[string]*schema.Schema{
 func ResourceKafkaSchemaConfiguration() *schema.Resource {
 	return &schema.Resource{
 		Description:   "The Kafka Schema Configuration resource allows the creation and management of Aiven Kafka Schema Configurations.",
-		CreateContext: resourceKafkaSchemaConfigurationCreate,
-		UpdateContext: resourceKafkaSchemaConfigurationUpdate,
-		ReadContext:   resourceKafkaSchemaConfigurationRead,
-		DeleteContext: resourceKafkaSchemaConfigurationDelete,
+		CreateContext: common.WithGenClientDiag(resourceKafkaSchemaConfigurationCreate),
+		UpdateContext: common.WithGenClientDiag(resourceKafkaSchemaConfigurationUpdate),
+		ReadContext:   common.WithGenClientDiag(resourceKafkaSchemaConfigurationRead),
+		DeleteContext: common.WithGenClientDiag(resourceKafkaSchemaConfigurationDelete),
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
 		},
@@ -46,37 +51,37 @@ func ResourceKafkaSchemaConfiguration() *schema.Resource {
 	}
 }
 
-func resourceKafkaSchemaConfigurationUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceKafkaSchemaConfigurationUpdate(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
 	project, serviceName, err := schemautil.SplitResourceID2(d.Id())
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
-	_, err = m.(*aiven.Client).KafkaGlobalSchemaConfig.Update(
+	_, err = client.ServiceSchemaRegistryGlobalConfigPut(
 		ctx,
 		project,
 		serviceName,
-		aiven.KafkaSchemaConfig{
-			CompatibilityLevel: d.Get("compatibility_level").(string),
+		&kafkaschemaregistry.ServiceSchemaRegistryGlobalConfigPutIn{
+			Compatibility: kafkaschemaregistry.CompatibilityType(d.Get("compatibility_level").(string)),
 		})
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
-	return resourceKafkaSchemaConfigurationRead(ctx, d, m)
+	return resourceKafkaSchemaConfigurationRead(ctx, d, client)
 }
 
 // resourceKafkaSchemaConfigurationCreate Kafka Schemas global configuration cannot be created but only updated
-func resourceKafkaSchemaConfigurationCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceKafkaSchemaConfigurationCreate(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
 	project := d.Get("project").(string)
 	serviceName := d.Get("service_name").(string)
 
-	_, err := m.(*aiven.Client).KafkaGlobalSchemaConfig.Update(
+	_, err := client.ServiceSchemaRegistryGlobalConfigPut(
 		ctx,
 		project,
 		serviceName,
-		aiven.KafkaSchemaConfig{
-			CompatibilityLevel: d.Get("compatibility_level").(string),
+		&kafkaschemaregistry.ServiceSchemaRegistryGlobalConfigPutIn{
+			Compatibility: kafkaschemaregistry.CompatibilityType(d.Get("compatibility_level").(string)),
 		})
 	if err != nil {
 		return diag.FromErr(err)
@@ -84,16 +89,26 @@ func resourceKafkaSchemaConfigurationCreate(ctx context.Context, d *schema.Resou
 
 	d.SetId(schemautil.BuildResourceID(project, serviceName))
 
-	return resourceKafkaSchemaConfigurationRead(ctx, d, m)
+	return resourceKafkaSchemaConfigurationRead(ctx, d, client)
 }
 
-func resourceKafkaSchemaConfigurationRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceKafkaSchemaConfigurationRead(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
 	project, serviceName, err := schemautil.SplitResourceID2(d.Id())
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
-	r, err := m.(*aiven.Client).KafkaGlobalSchemaConfig.Get(ctx, project, serviceName)
+	// if Schema Registry is disabled, the resource is considered deleted
+	enabled, err := isSchemaRegistryEnabled(ctx, client, project, serviceName)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	if !enabled {
+		d.SetId("")
+		return nil
+	}
+
+	compatibilityLevel, err := client.ServiceSchemaRegistryGlobalConfigGet(ctx, project, serviceName)
 	if err != nil {
 		return diag.FromErr(schemautil.ResourceReadHandleNotFound(err, d))
 	}
@@ -104,7 +119,7 @@ func resourceKafkaSchemaConfigurationRead(ctx context.Context, d *schema.Resourc
 	if err := d.Set("service_name", serviceName); err != nil {
 		return diag.FromErr(err)
 	}
-	if err := d.Set("compatibility_level", r.CompatibilityLevel); err != nil {
+	if err := d.Set("compatibility_level", string(compatibilityLevel)); err != nil {
 		return diag.FromErr(err)
 	}
 
@@ -113,22 +128,56 @@ func resourceKafkaSchemaConfigurationRead(ctx context.Context, d *schema.Resourc
 
 // resourceKafkaSchemaConfigurationDelete Kafka Schemas configuration cannot be deleted, therefore
 // on delete event configuration will be set to the default setting
-func resourceKafkaSchemaConfigurationDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+func resourceKafkaSchemaConfigurationDelete(ctx context.Context, d *schema.ResourceData, client avngen.Client) diag.Diagnostics {
 	project, serviceName, err := schemautil.SplitResourceID2(d.Id())
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
-	_, err = m.(*aiven.Client).KafkaGlobalSchemaConfig.Update(
+	enabled, err := isSchemaRegistryEnabled(ctx, client, project, serviceName)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	if !enabled {
+		return nil
+	}
+
+	_, err = client.ServiceSchemaRegistryGlobalConfigPut(
 		ctx,
 		project,
 		serviceName,
-		aiven.KafkaSchemaConfig{
-			CompatibilityLevel: "BACKWARD",
+		&kafkaschemaregistry.ServiceSchemaRegistryGlobalConfigPutIn{
+			Compatibility: kafkaschemaregistry.CompatibilityTypeBackward,
 		})
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
 	return nil
+}
+
+// isSchemaRegistryEnabled checks if the Schema Registry is enabled
+func isSchemaRegistryEnabled(ctx context.Context, client avngen.Client, project, serviceName string) (bool, error) {
+	key := filepath.Join(project, serviceName)
+
+	v, err, _ := schemaRegistryCheckGroup.Do(key, func() (any, error) {
+		service, err := client.ServiceGet(ctx, project, serviceName)
+		if err != nil {
+			return false, err
+		}
+
+		if v, ok := service.UserConfig["schema_registry"]; ok {
+			if val, ok := v.(bool); ok {
+				return val, nil
+			}
+		}
+
+		return false, nil
+	})
+
+	if err != nil {
+		return false, err
+	}
+
+	return v.(bool), nil
 }